Change code according to comment.
