+++ flink-connectors/flink-connector-jdbc/src/main/java/org/apache/flink/connector/jdbc/internal/options/JdbcOptions.java
@@ -34,7 +34,7 @@ public class JdbcOptions extends JdbcConnectionOptions {
 
 	private static final long serialVersionUID = 1L;
 
-	public static final int CONNECTION_CHECK_TIMEOUT = 60;
+	public static final int CONNECTION_CHECK_TIMEOUT_SECONDS = 60;
 
 	private String tableName;
 	private JdbcDialect dialect;
