+++ flink-connectors/flink-connector-jdbc/src/test/java/org/apache/flink/connector/jdbc/table/JdbcLookupTableITCase.java
@@ -19,7 +19,6 @@
 package org.apache.flink.connector.jdbc.table;
 
 import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.connector.jdbc.JdbcTestFixture;
 import org.apache.flink.connector.jdbc.internal.options.JdbcLookupOptions;
 import org.apache.flink.connector.jdbc.internal.options.JdbcOptions;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
@@ -28,21 +27,14 @@ import org.apache.flink.table.api.Table;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
 import org.apache.flink.table.types.DataType;
-import org.apache.flink.test.util.AbstractTestBase;
 import org.apache.flink.types.Row;
 
 import org.apache.flink.shaded.guava18.com.google.common.collect.Lists;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -51,7 +43,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static org.apache.flink.connector.jdbc.JdbcTestFixture.DERBY_EBOOKSHOP_DB;
 import static org.apache.flink.table.api.Expressions.$;
 import static org.junit.Assert.assertEquals;
 
@@ -59,10 +50,7 @@ import static org.junit.Assert.assertEquals;
  * IT case for lookup source of JDBC connector.
  */
 @RunWith(Parameterized.class)
-public class JdbcLookupTableITCase extends AbstractTestBase {
-
-	public static final String DB_URL = "jdbc:derby:memory:lookup";
-	public static final String LOOKUP_TABLE = "lookup_table";
+public class JdbcLookupTableITCase extends JdbcLookupTestBase {
 
 	private final String tableFactory;
 	private final boolean useCache;
@@ -80,70 +68,6 @@ public class JdbcLookupTableITCase extends AbstractTestBase {
 			{"legacyFactory", false},
 			{"dynamicFactory", true},
 			{"dynamicFactory", false}});
-	}
-
-	@Before
-	public void before() throws ClassNotFoundException, SQLException {
-		System.setProperty("derby.stream.error.field", JdbcTestFixture.class.getCanonicalName() + ".DEV_NULL");
-
-		Class.forName(DERBY_EBOOKSHOP_DB.getDriverClass());
-		try (
-			Connection conn = DriverManager.getConnection(DB_URL + ";create=true");
-			Statement stat = conn.createStatement()) {
-			stat.executeUpdate("CREATE TABLE " + LOOKUP_TABLE + " (" +
-					"id1 INT NOT NULL DEFAULT 0," +
-					"id2 VARCHAR(20) NOT NULL," +
-					"comment1 VARCHAR(1000)," +
-					"comment2 VARCHAR(1000))");
-
-			Object[][] data = new Object[][] {
-					new Object[] {1, "1", "11-c1-v1", "11-c2-v1"},
-					new Object[] {1, "1", "11-c1-v2", "11-c2-v2"},
-					new Object[] {2, "3", null, "23-c2"},
-					new Object[] {2, "5", "25-c1", "25-c2"},
-					new Object[] {3, "8", "38-c1", "38-c2"}
-			};
-			boolean[] surroundedByQuotes = new boolean[] {
-				false, true, true, true
-			};
-
-			StringBuilder sqlQueryBuilder = new StringBuilder(
-					"INSERT INTO " + LOOKUP_TABLE + " (id1, id2, comment1, comment2) VALUES ");
-			for (int i = 0; i < data.length; i++) {
-				sqlQueryBuilder.append("(");
-				for (int j = 0; j < data[i].length; j++) {
-					if (data[i][j] == null) {
-						sqlQueryBuilder.append("null");
-					} else {
-						if (surroundedByQuotes[j]) {
-							sqlQueryBuilder.append("'");
-						}
-						sqlQueryBuilder.append(data[i][j]);
-						if (surroundedByQuotes[j]) {
-							sqlQueryBuilder.append("'");
-						}
-					}
-					if (j < data[i].length - 1) {
-						sqlQueryBuilder.append(", ");
-					}
-				}
-				sqlQueryBuilder.append(")");
-				if (i < data.length - 1) {
-					sqlQueryBuilder.append(", ");
-				}
-			}
-			stat.execute(sqlQueryBuilder.toString());
-		}
-	}
-
-	@After
-	public void clearOutputTable() throws Exception {
-		Class.forName(DERBY_EBOOKSHOP_DB.getDriverClass());
-		try (
-				Connection conn = DriverManager.getConnection(DB_URL);
-				Statement stat = conn.createStatement()) {
-			stat.execute("DROP TABLE " + LOOKUP_TABLE);
-		}
 	}
 
 	@Test
