+++ flink-connectors/flink-connector-jdbc/src/test/java/org/apache/flink/connector/jdbc/table/JdbcRowDataLookupFunctionTest.java
@@ -18,7 +18,6 @@
 
 package org.apache.flink.connector.jdbc.table;
 
-import org.apache.flink.connector.jdbc.JdbcTestFixture;
 import org.apache.flink.connector.jdbc.internal.options.JdbcLookupOptions;
 import org.apache.flink.connector.jdbc.internal.options.JdbcOptions;
 import org.apache.flink.table.api.DataTypes;
@@ -27,35 +26,25 @@ import org.apache.flink.table.data.StringData;
 import org.apache.flink.table.types.DataType;
 import org.apache.flink.table.types.logical.LogicalType;
 import org.apache.flink.table.types.logical.RowType;
-import org.apache.flink.test.util.AbstractTestBase;
 import org.apache.flink.util.Collector;
 
 import org.apache.flink.shaded.guava18.com.google.common.collect.Lists;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static org.apache.flink.connector.jdbc.JdbcTestFixture.DERBY_EBOOKSHOP_DB;
 import static org.junit.Assert.assertEquals;
 
 /**
  * Test suite for {@link JdbcRowDataLookupFunction}.
  */
-public class JdbcRowDataLookupFunctionTest extends AbstractTestBase {
-
-	public static final String DB_URL = "jdbc:derby:memory:lookup";
-	public static final String LOOKUP_TABLE = "lookup_table";
-	public static final String DB_DRIVER = "org.apache.derby.jdbc.EmbeddedDriver";
+public class JdbcRowDataLookupFunctionTest extends JdbcLookupTestBase {
 
 	private static String[] fieldNames = new String[] {"id1", "id2", "comment1", "comment2"};
 	private static DataType[] fieldDataTypes = new DataType[] {
@@ -67,132 +56,40 @@ public class JdbcRowDataLookupFunctionTest extends AbstractTestBase {
 
 	private static String[] lookupKeys = new String[] {"id1", "id2"};
 
-	@Before
-	public void before() throws ClassNotFoundException, SQLException {
-		System.setProperty("derby.stream.error.field", JdbcTestFixture.class.getCanonicalName() + ".DEV_NULL");
-
-		Class.forName(DB_DRIVER);
-		try (
-			Connection conn = DriverManager.getConnection(DB_URL + ";create=true");
-			Statement stat = conn.createStatement()) {
-			stat.executeUpdate("CREATE TABLE " + LOOKUP_TABLE + " (" +
-				"id1 INT NOT NULL DEFAULT 0," +
-				"id2 VARCHAR(20) NOT NULL," +
-				"comment1 VARCHAR(1000)," +
-				"comment2 VARCHAR(1000))");
-
-			Object[][] data = new Object[][] {
-				new Object[] {1, "1", "11-c1-v1", "11-c2-v1"},
-				new Object[] {1, "1", "11-c1-v2", "11-c2-v2"},
-				new Object[] {2, "3", null, "23-c2"},
-				new Object[] {2, "5", "25-c1", "25-c2"},
-				new Object[] {3, "8", "38-c1", "38-c2"}
-			};
-			boolean[] surroundedByQuotes = new boolean[] {
-				false, true, true, true
-			};
-
-			StringBuilder sqlQueryBuilder = new StringBuilder(
-				"INSERT INTO " + LOOKUP_TABLE + " (id1, id2, comment1, comment2) VALUES ");
-			for (int i = 0; i < data.length; i++) {
-				sqlQueryBuilder.append("(");
-				for (int j = 0; j < data[i].length; j++) {
-					if (data[i][j] == null) {
-						sqlQueryBuilder.append("null");
-					} else {
-						if (surroundedByQuotes[j]) {
-							sqlQueryBuilder.append("'");
-						}
-						sqlQueryBuilder.append(data[i][j]);
-						if (surroundedByQuotes[j]) {
-							sqlQueryBuilder.append("'");
-						}
-					}
-					if (j < data[i].length - 1) {
-						sqlQueryBuilder.append(", ");
-					}
-				}
-				sqlQueryBuilder.append(")");
-				if (i < data.length - 1) {
-					sqlQueryBuilder.append(", ");
-				}
-			}
-			stat.execute(sqlQueryBuilder.toString());
-		}
-	}
-
-	@After
-	public void clearOutputTable() throws Exception {
-		Class.forName(DB_DRIVER);
-		try (
-			Connection conn = DriverManager.getConnection(DB_URL);
-			Statement stat = conn.createStatement()) {
-			stat.execute("DROP TABLE " + LOOKUP_TABLE);
-		}
-	}
-
 	@Test
 	public void testEval() throws Exception {
 
 		JdbcRowDataLookupFunction lookupFunction = buildRowDataLookupFunction();
 
-		lookupFunction.open(null);
-
 		ListOutputCollector collector = new ListOutputCollector();
 		lookupFunction.setCollector(collector);
 
-		lookupFunction.eval(1, StringData.fromString("1"));
-
-		List<String> result = Lists.newArrayList(collector.getOutputs()).stream()
-			.map(row -> {
-				return String.format("%d,%s,%s,%s",
-					row.getInt(0), row.getString(1), row.getString(2), row.getString(3));
-			})
-			.sorted()
-			.collect(Collectors.toList());
-
-		List<String> expected = new ArrayList<>();
-		expected.add("1,1,11-c1-v1,11-c2-v1");
-		expected.add("1,1,11-c1-v2,11-c2-v2");
-		Collections.sort(expected);
-
-		assertEquals(expected, result);
-	}
-
-	@Test
-	public void testInvalidConnectionInEval() throws Exception {
-
-		JdbcRowDataLookupFunction lookupFunction = buildRowDataLookupFunction();
-
 		lookupFunction.open(null);
+
+		lookupFunction.eval(1, StringData.fromString("1"));
 
 		// close connection
 		lookupFunction.getDbConnection().close();
 
-		ListOutputCollector collector = new ListOutputCollector();
-		lookupFunction.setCollector(collector);
-
-		lookupFunction.eval(1, StringData.fromString("1"));
+		lookupFunction.eval(2, StringData.fromString("3"));
 
 		List<String> result = Lists.newArrayList(collector.getOutputs()).stream()
-			.map(row -> {
-				return String.format("%d,%s,%s,%s",
-					row.getInt(0), row.getString(1), row.getString(2), row.getString(3));
-			})
+			.map(RowData::toString)
 			.sorted()
 			.collect(Collectors.toList());
 
 		List<String> expected = new ArrayList<>();
-		expected.add("1,1,11-c1-v1,11-c2-v1");
-		expected.add("1,1,11-c1-v2,11-c2-v2");
+		expected.add("+I(1,1,11-c1-v1,11-c2-v1)");
+		expected.add("+I(1,1,11-c1-v2,11-c2-v2)");
+		expected.add("+I(2,3,null,23-c2)");
 		Collections.sort(expected);
 
 		assertEquals(expected, result);
 	}
 
-	JdbcRowDataLookupFunction buildRowDataLookupFunction() {
+	private JdbcRowDataLookupFunction buildRowDataLookupFunction() {
 		JdbcOptions jdbcOptions = JdbcOptions.builder()
-			.setDriverName(DB_DRIVER)
+			.setDriverName(DERBY_EBOOKSHOP_DB.getDriverClass())
 			.setDBUrl(DB_URL)
 			.setTableName(LOOKUP_TABLE)
 			.build();
